Share MyBean3Test and pin down where a TYPE_USE annotation lands
The two copies disagreed on one assertion: the source suite looked for
jspecify's @Nullable on the parameter's annotated type, the ASM suite on the
parameter itself. jspecify's @Nullable is @Target(TYPE_USE) only, so the
source suite is right and the bytecode backends write it as a plain
parameter annotation instead of a type annotation.

Keep one copy in the TCK asserting the correct placement, and split that
assertion into its own test so the rest of the class keeps running on every
backend rather than the whole test being lost to the difference. The new
test is disabled for the bytecode backends with the reason stated.

Also drop the unused Disabled import the source suite's copy carried.

Every backend gains a test; none is lost.
