+++ test-suite-tck/src/test/java/io/micronaut/sourcegen/example/MyBean3Test.java
@@ -17,6 +17,7 @@ package io.micronaut.sourcegen.example;
 
 import org.jspecify.annotations.Nullable;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
 
 import java.io.IOException;
 import java.lang.reflect.Constructor;
@@ -39,14 +40,32 @@ class MyBean3Test {
         assertNotNull(constructor);
         Parameter parameter = constructor.getParameters()[0];
         assertNotNull(parameter);
-        assertNotNull(parameter.getDeclaredAnnotation(Nullable.class));
     }
 
     @Test
-    void testConcatenate() throws Exception {
+    @DisabledIfSystemProperty(named = "sourcegen.backend", matches = "bytecode.*",
+        disabledReason = "The bytecode backends write a TYPE_USE-only annotation as a plain "
+            + "parameter annotation, so it is not on the parameter's annotated type")
+    void nullableOnAParameterIsWrittenAsATypeAnnotation() throws Exception {
+        // org.jspecify.annotations.Nullable is @Target(TYPE_USE) only, so it belongs to the
+        // parameter's type rather than to the parameter itself
+        Constructor<MyBean3> constructor = MyBean3.class.getDeclaredConstructor(Integer.class);
+        Parameter parameter = constructor.getParameters()[0];
+        assertNotNull(parameter.getAnnotatedType().getDeclaredAnnotation(Nullable.class));
+    }
+
+    @Test
+    void testConcatenate() {
         MyBean3 bean3 = new MyBean3();
 
         assertEquals("Hello, Andriy", bean3.concatenation("Andriy"));
+    }
+
+    @Test
+    void testThrows() {
+        MyBean3 bean3 = new MyBean3();
+
+        assertThrows(IOException.class, bean3::getStringUnsafe);
     }
 
     @Test
@@ -54,12 +73,5 @@ class MyBean3Test {
         MyBean3 bean3 = new MyBean3();
 
         assertEquals("Count: 3, price: 1.5, flag: true", bean3.concatenationWithPrimitives(3, 1.5, true));
-    }
-
-    @Test
-    void testThrows() throws Exception {
-        MyBean3 bean3 = new MyBean3();
-
-        assertThrows(IOException.class, bean3::getStringUnsafe);
     }
 }
